+++ ompi/mca/io/ompio/io_ompio_file_write.c
@@ -12,6 +12,8 @@
  * Copyright (c) 2008-2019 University of Houston. All rights reserved.
  * Copyright (c) 2015-2018 Research Organization for Information Science
  *                         and Technology (RIST). All rights reserved.
+ * Copyright (c) 2024      Triad National Security, LLC. All rights
+ *                         reserved.
  * $COPYRIGHT$
  *
  * Additional copyrights may follow
@@ -56,7 +58,7 @@
 
 int mca_io_ompio_file_write (ompi_file_t *fp,
 			     const void *buf,
-			     int count,
+			     size_t count,
 			     struct ompi_datatype_t *datatype,
 			     ompi_status_public_t *status)
 {
@@ -76,7 +78,7 @@ int mca_io_ompio_file_write (ompi_file_t *fp,
 int mca_io_ompio_file_write_at (ompi_file_t *fh,
 				OMPI_MPI_OFFSET_TYPE offset,
 				const void *buf,
-				int count,
+				size_t count,
 				struct ompi_datatype_t *datatype,
 				ompi_status_public_t *status)
 {
@@ -93,7 +95,7 @@ int mca_io_ompio_file_write_at (ompi_file_t *fh,
 
 int mca_io_ompio_file_iwrite (ompi_file_t *fp,
 			      const void *buf,
-			      int count,
+			      size_t count,
 			      struct ompi_datatype_t *datatype,
 			      ompi_request_t **request)
 {
@@ -112,7 +114,7 @@ int mca_io_ompio_file_iwrite (ompi_file_t *fp,
 int mca_io_ompio_file_iwrite_at (ompi_file_t *fh,
 				 OMPI_MPI_OFFSET_TYPE offset,
 				 const void *buf,
-				 int count,
+				 size_t count,
 				 struct ompi_datatype_t *datatype,
 				 ompi_request_t **request)
 {
@@ -133,7 +135,7 @@ int mca_io_ompio_file_iwrite_at (ompi_file_t *fh,
 
 int mca_io_ompio_file_write_all (ompi_file_t *fh,
 				 const void *buf,
-				 int count,
+				 size_t count,
 				 struct ompi_datatype_t *datatype,
 				 ompi_status_public_t *status)
 {
@@ -162,7 +164,7 @@ int mca_io_ompio_file_write_all (ompi_file_t *fh,
 int mca_io_ompio_file_write_at_all (ompi_file_t *fh,
 				    OMPI_MPI_OFFSET_TYPE offset,
 				    const void *buf,
-				    int count,
+				    size_t count,
 				    struct ompi_datatype_t *datatype,
 				    ompi_status_public_t *status)
 {
@@ -179,7 +181,7 @@ int mca_io_ompio_file_write_at_all (ompi_file_t *fh,
 
 int mca_io_ompio_file_iwrite_all (ompi_file_t *fh,
 				  const void *buf,
-				  int count,
+				  size_t count,
 				  struct ompi_datatype_t *datatype,
 				  ompi_request_t **request)
 {
@@ -203,7 +205,7 @@ int mca_io_ompio_file_iwrite_all (ompi_file_t *fh,
 int mca_io_ompio_file_iwrite_at_all (ompi_file_t *fh,
 				     OMPI_MPI_OFFSET_TYPE offset,
 				     const void *buf,
-				     int count,
+				     size_t count,
 				     struct ompi_datatype_t *datatype,
 				     ompi_request_t **request)
 {
@@ -224,7 +226,7 @@ int mca_io_ompio_file_iwrite_at_all (ompi_file_t *fh,
 
 int mca_io_ompio_file_write_shared (ompi_file_t *fp,
 				    const void *buf,
-				    int count,
+				    size_t count,
 				    struct ompi_datatype_t *datatype,
 				    ompi_status_public_t * status)
 {
@@ -251,7 +253,7 @@ int mca_io_ompio_file_write_shared (ompi_file_t *fp,
 
 int mca_io_ompio_file_iwrite_shared (ompi_file_t *fp,
 				     const void *buf,
-				     int count,
+				     size_t count,
 				     struct ompi_datatype_t *datatype,
 				     ompi_request_t **request)
 {
@@ -278,7 +280,7 @@ int mca_io_ompio_file_iwrite_shared (ompi_file_t *fp,
 
 int mca_io_ompio_file_write_ordered (ompi_file_t *fp,
 				     const void *buf,
-				     int count,
+				     size_t count,
 				     struct ompi_datatype_t *datatype,
 				     ompi_status_public_t * status)
 {
@@ -305,7 +307,7 @@ int mca_io_ompio_file_write_ordered (ompi_file_t *fp,
 
 int mca_io_ompio_file_write_ordered_begin (ompi_file_t *fp,
 					   const void *buf,
-					   int count,
+					   size_t count,
 					   struct ompi_datatype_t *datatype)
 {
     int ret = OMPI_SUCCESS;
@@ -359,7 +361,7 @@ int mca_io_ompio_file_write_ordered_end (ompi_file_t *fp,
 /**********************************************************************/
 int mca_io_ompio_file_write_all_begin (ompi_file_t *fh,
 				       const void *buf,
-				       int count,
+				       size_t count,
 				       struct ompi_datatype_t *datatype)
 {
     int ret = OMPI_SUCCESS;
@@ -400,7 +402,7 @@ int mca_io_ompio_file_write_all_end (ompi_file_t *fh,
 int mca_io_ompio_file_write_at_all_begin (ompi_file_t *fh,
 					  OMPI_MPI_OFFSET_TYPE offset,
 					  const void *buf,
-					  int count,
+					  size_t count,
 					  struct ompi_datatype_t *datatype)
 {
     int ret = OMPI_SUCCESS;
